requirements.txt
keras==2.4.0
numpy
pillow
six==1.15.0
tensorflow==2.4.0
tensorflow-gpu==2.4.0
tensorflow_addons
tqdm
opencv-python
colorthief
matplotlib